examples/requirements.txt
matplotlib>=3.10.1
numpy>=2.2.4
opencv_python>=4.9.0.80
Pillow>=11.1.0
scipy>=1.15.2
torch>=2.6.0
torchvision>=0.21.0
